trim whitespace from guile-config output
Previously, the last argument from guile-config would include the
trailing newline. This gets passed to clang with the newline and doesn't
work as expected.

On my computer, `guile-config compile` returns:

    -pthread -I/usr/include/guile/2.2

which causes the string "-I/usr/include/guile/2.2\n" to get passed to
libclang, which prevents bindgen from finding the guile header files.

I fixed this by adding .trim() to remove the whitespace from the
beginning and end of the guile-config output.